alfheim base

+++ src/main/java/com/Nxer/world/TestUseOnly.java
@@ -1,5 +1,6 @@
 package com.Nxer.world;
 
 public class TestUseOnly {
-    public static int[] skystats = new int[]{0,0,0,0,0,0}; 
+
+    public static float[] skystats = new float[] { 0F, 0F, 0F, 0F, 0F, 0F };
 }
